🏗️ Move `InvitForm` component into new dedicated feature folder
