+++ src/features/admin/invitation/components/InvitForm/style.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/breakpoints' as b;
 
 .c-invit-form {
